error on bad type annotation at def time
