+++ python_modules/dagster/dagster/_core/definitions/decorators/asset_decorator.py
@@ -21,7 +21,7 @@ import dagster._check as check
 from dagster._annotations import deprecated_param, experimental_param
 from dagster._builtins import Nothing
 from dagster._config import UserConfigSchema
-from dagster._core.decorator_utils import get_function_params, get_valid_name_permutations
+from dagster._core.decorator_utils import get_function_params
 from dagster._core.definitions.asset_dep import AssetDep, CoercibleToAssetDep
 from dagster._core.definitions.auto_materialize_policy import AutoMaterializePolicy
 from dagster._core.definitions.config import ConfigMapping
@@ -854,11 +854,10 @@ def multi_asset(
 
 
 def get_function_params_without_context_or_config_or_resources(fn: Callable) -> List[Parameter]:
+    from dagster._core.decorator_utils import is_context_provided
+
     params = get_function_params(fn)
-    is_context_provided = len(params) > 0 and params[0].name in get_valid_name_permutations(
-        "context"
-    )
-    input_params = params[1:] if is_context_provided else params
+    input_params = params[1:] if is_context_provided(params) else params
 
     resource_arg_names = {arg.name for arg in get_resource_args(fn)}
 
@@ -1341,3 +1340,18 @@ def _get_partition_mappings_from_deps(
             )
 
     return partition_mappings
+
+
+def _validate_context_type_hint(fn):
+    from inspect import _empty as EmptyAnnotation
+
+    from dagster._core.decorator_utils import get_function_params, is_context_provided
+    from dagster._core.execution.context.compute import AssetExecutionContext, OpExecutionContext
+
+    params = get_function_params(fn)
+    if is_context_provided(params):
+        if not isinstance(params[0], (AssetExecutionContext, OpExecutionContext, EmptyAnnotation)):
+            raise DagsterInvalidDefinitionError(
+                f"Cannot annotate `context` parameter with type {params[0].annotation}. `context`"
+                " must be annotated with AssetExecutionContext, OpExecutionContext, or left blank."
+            )
